Use stdint.h to define all integers, stdbool.h for booleans
This assures maximum portability and performance as the integer values
are exactly as small as they could be. At the same time the easy usage
of unsigned values makes the values impossible to hold negative values
when they should not (such as the twelfths from 0 to 11).

The stdbool.h was already used by Numerus only internally, while the
outputs of the functions returning booleans were short integers.
Including stdbool.h into the numerus.h header file allows to use this
macros also for the public user of the library. This has been done for
code clarity as "true" and "false" are more understandable than a 0
and 1, while still being a 0 and 1 if needed.

+++ src/numerus_cli.c
@@ -157,7 +157,7 @@ static short _num_string_is_double_zero(char *zero_as_string) {
 void _num_convert_to_other_form_and_print(char *string) {
     double value;
     char *roman;
-    int errcode;
+    int8_t errcode;
     /* This check is necessary because strtod() returns 0 in case of errors
      * AND in case finds an actual zero. Duh! */
     if (_num_string_is_double_zero(string)) {
@@ -282,7 +282,7 @@ static int _num_parse_command(char *command) {
  * @returns int status code: 0 if everything went OK or a NUMERUS_ERROR_*
  * otherwise.
  */
-int numerus_cli(int argc, char **args) {
+int8_t numerus_cli(int argc, char **args) {
     char *command;
     /* line_buffer_size = 50 enough for every command,
      * gets reallocated by getline() if not enough */
